refactor: rename option-list to option
