+++ src/framework/theme/components/option/_option-list.component.theme.scss
