+++ src/framework/theme/components/option/option-group.component.spec.ts
