+++ src/framework/theme/components/option/option.component.scss
