+++ src/framework/theme/public_api.ts
@@ -164,10 +164,10 @@ export * from './components/tooltip/tooltip.directive';
 export * from './components/tooltip/tooltip.component';
 export * from './components/select/select.module';
 export * from './components/select/select.component';
-export * from './components/option-list/option-list.module';
-export * from './components/option-list/option.component';
-export * from './components/option-list/option-group.component';
-export * from './components/option-list/option-list.component';
+export * from './components/option/option-list.module';
+export * from './components/option/option.component';
+export * from './components/option/option-group.component';
+export * from './components/option/option-list.component';
 export * from './components/select/select-injection-tokens';
 export * from './components/autocomplete/autocomplete.module';
 export * from './components/autocomplete/autocomplete.component';
