+++ src/framework/theme/styles/global/_components.scss
@@ -33,7 +33,7 @@
 @import '../../components/popover/popover.component.theme';
 @import '../../components/context-menu/context-menu.component.theme';
 @import '../../components/select/select.component.theme';
-@import '../../components/option-list/option-list.component.theme';
+@import '../../components/option/option-list.component.theme';
 @import '../../components/toastr/toast.component.theme';
 @import '../../components/tooltip/tooltip.component.theme';
 @import '../../components/datepicker/datepicker-container.component.theme';
